v2.0.3 - implemented add validator functionality - added remaining stake in validator list

+++ lib/src/api/apiNomination.dart
@@ -16,7 +16,7 @@ class ApiNomination {
     required String nodeID,
     required String amount,
     required int end,
-    required int fee,
+    required double fee,
     String? rewardAddress,
   }) async {
     final res = await service.addValidator(
