+++ lib/src/service/nomination.dart
@@ -15,7 +15,7 @@ class ServiceNomination {
     required String nodeID,
     required String amount,
     required int end,
-    required int fee,
+    required double fee,
     String? rewardAddress,
   }) async {
     DateTime endDate = DateTime.fromMillisecondsSinceEpoch(end);
